Default prompt-cache TTL to 5m; 1h premium wasn't amortizing
Measured cacheWrite:cacheRead ratio on a $45 project run sat at 0.84:1 —
writes exceeded reads in total tokens, meaning many cached prefixes never
got read back before the next write invalidated them. The 1h TTL was
paying a 60% premium on every write without buying extra reuse: only 3
gaps between brain turns exceeded 5m (max 11m), and 0 exceeded 1h.

At log-observed scaling, switching to 5m saves ~$14.77 on a $45 run
(~33%) while keeping the same cache hit pattern. Workloads with proven
long idle gaps can opt back in via PI_CACHE_RETENTION=long.

Updated both src/agent.ts and src/subagent-runner.ts since sub-agents run
in separate processes and need the env var set independently.

+++ src/agent.ts
@@ -41,11 +41,14 @@ import { installUsageTracking, readUsageTotals } from "./usage-log.js";
 import type { ThinkingLevel } from "@mariozechner/pi-agent-core";
 import type { PIVerdict } from "./pi-agent.js";
 
-// Default Anthropic prompt-cache TTL to 1h. Luxas runs span hours with idle
-// gaps between agents — the 5-min default expires mid-run and forces full
-// re-caching on every cold spawn. 1h costs 2× per write but is amortized
-// over 12× more reuse. Override with PI_CACHE_RETENTION=short.
-process.env.PI_CACHE_RETENTION ||= "long";
+// Default Anthropic prompt-cache TTL to 5m. Empirically on Luxas runs the
+// cacheWrite/cacheRead ratio sits near 1:1 (mutable L3 + shifting context
+// snapshots force frequent re-writes), so the 1h premium rarely amortizes —
+// large writes with few reads overpay for the extended TTL. Gaps between
+// brain turns are almost always < 5m because foreground spawns resume
+// quickly. Override with PI_CACHE_RETENTION=long for workloads with proven
+// long idle gaps (e.g. many >15min brain-side gaps).
+process.env.PI_CACHE_RETENTION ||= "short";
 
 export interface ResearchAgentOptions {
   projectDir: string;
